+++ src/hazardous/kem/mod.rs
@@ -20,8 +20,6 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-#[cfg(feature = "safe_api")]
-#[cfg_attr(docsrs, doc(cfg(feature = "safe_api")))]
 /// DHKEM(X25519, HKDF-SHA256) as specified in HPKE [RFC 9180](https://www.rfc-editor.org/rfc/rfc9180.html).
 pub mod x25519_hkdf_sha256;